credit-card-fraud-recognition
The PwC global economic crime survey of 2016 suggests that approximately 36% of organizations experienced economic crime. Therefore, there is deﬁnitely a need to solve the problem of credit card fraud detection. The task of fraud detection often boils down to outlier detection, in which a dataset is scanned through to ﬁnd potential anomalies in the data. In the past, this was done by employees which checked all transactions manually. A lot of research has been done in order to ﬁnd a solution to this problem. With the rise of machine learning, artiﬁcial intelligence, deep learning and other relevant ﬁelds of information technology, it becomes feasible to automate this process and to save some of the intensive amount of labor that is put into detecting credit card fraud.As already mentioned already that, with the help of machine learning techniques labor used in identifying the fraud transactions can be reduced to a great extent. Basically, the problem can be stated as a binary classiﬁcation i.e. fraud transaction or genuine transaction. But the frequencies of the two classes is very unbalanced in this case, so, we don’t have comparable number of observations for each classes. As part of this project, my aim is to create few machine learning models which can identify the fraud transactions from given data of transactions